Relocate asset rename executor into identifier-case
